Footer global + RX 0.00 + fixes formularios
Backend: Prescription usa BigDecimal (2 decimales) y migración V11 a DECIMAL(6,2); EJE a DECIMAL(5,2).
Frontend: ESF/CIL/EJE/ADD formateados a 0.00 en SPA y vistas estáticas.
RxPicker: EJE muestra 2 decimales; normalización/validación consistente.
Pacientes (Nuevo/Editar): Dirección y Comuna lado a lado; sin pins; Comuna preselecciona; duplicado removido en Ingresar.
Footer: global, tema‑aware, al final del contenido (sticky layout flex), visible en todas las vistas (incl. login); z‑index por encima del overlay; GitHub + Instagram + texto legal.

+++ src/main/java/cl/letelier/letelier/prescription/PrescriptionDTO.java
@@ -1,10 +1,11 @@
 package cl.letelier.letelier.prescription;
 
 import java.time.LocalDate;
+import java.math.BigDecimal;
 
 public record PrescriptionDTO(
     Long id, Long pacienteId,
-    Double odEsfera, Double odCilindro, Integer odEje,
-    Double oiEsfera, Double oiCilindro, Integer oiEje,
-    Double addPower, String observaciones, LocalDate fecha, boolean activo
+    BigDecimal odEsfera, BigDecimal odCilindro, BigDecimal odEje,
+    BigDecimal oiEsfera, BigDecimal oiCilindro, BigDecimal oiEje,
+    BigDecimal addPower, String observaciones, LocalDate fecha, boolean activo
 ){}
